Revert "Fixed types for future proofing of default export."
This reverts commit 6f44519fb804de6d4523094a199f4486a350e209.
